LST-1162 Fix floor availability check not actually being atomic.
The old code forgot to account for the fact Django lazily evaluates
querysets, so it never actually ran the "SELECT ... FOR UPDATE" query.

Use .get() instead to force Django to execute the query and lock the
floor.

The other change is just re-formatting the code a bit to minimize the
amount of time we hold a lock on the DB row, no functional change.